backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
pyodbc>=5.0
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0
cryptography>=41.0
